move delegates to package
