+++ engine/src/main/java/ai/hhrdr/chainflow/engine/delegate/AddEventDelegate.java
@@ -1,4 +1,4 @@
-package ai.hhrdr.chainflow.engine;
+package ai.hhrdr.chainflow.engine.delegate;
 
 import org.camunda.bpm.engine.delegate.DelegateExecution;
 import org.camunda.bpm.engine.delegate.JavaDelegate;
