+++ engine/src/main/resources/invite_season_2_pass_start_form.bpmn
@@ -53,7 +53,7 @@
         <bpmn:flowNodeRef>Event_1uuift7</bpmn:flowNodeRef>
       </bpmn:lane>
     </bpmn:laneSet>
-    <bpmn:serviceTask id="Activity_1golbhl" name="Guru Network Transport Layer" camunda:class="ai.hhrdr.chainflow.engine.LoggerDelegate">
+    <bpmn:serviceTask id="Activity_1golbhl" name="Guru Network Transport Layer" camunda:class="ai.hhrdr.chainflow.engine.delegate.LoggerDelegate">
       <bpmn:incoming>Flow_03gn7ax</bpmn:incoming>
       <bpmn:outgoing>Flow_0iu3jen</bpmn:outgoing>
     </bpmn:serviceTask>
@@ -152,7 +152,7 @@
       </bpmn:extensionElements>
       <bpmn:outgoing>Flow_00l3m44</bpmn:outgoing>
     </bpmn:startEvent>
-    <bpmn:serviceTask id="Activity_1514wn9" name="Set ENV Vars and Calculate flags based on provided inputs" camunda:class="ai.hhrdr.chainflow.engine.LoggerDelegate">
+    <bpmn:serviceTask id="Activity_1514wn9" name="Set ENV Vars and Calculate flags based on provided inputs" camunda:class="ai.hhrdr.chainflow.engine.delegate.LoggerDelegate">
       <bpmn:incoming>Flow_00l3m44</bpmn:incoming>
       <bpmn:outgoing>Flow_03gn7ax</bpmn:outgoing>
     </bpmn:serviceTask>
@@ -169,7 +169,7 @@
       <bpmn:incoming>Flow_0iu3jen</bpmn:incoming>
       <bpmn:outgoing>Flow_0wtr19b</bpmn:outgoing>
     </bpmn:callActivity>
-    <bpmn:serviceTask id="Activity_03au33a" name="Guru Network Transport Layer" camunda:class="ai.hhrdr.chainflow.engine.LoggerDelegate">
+    <bpmn:serviceTask id="Activity_03au33a" name="Guru Network Transport Layer" camunda:class="ai.hhrdr.chainflow.engine.delegate.LoggerDelegate">
       <bpmn:incoming>Flow_0wtr19b</bpmn:incoming>
       <bpmn:outgoing>Flow_1344k37</bpmn:outgoing>
     </bpmn:serviceTask>
